Performance improvements and added code to model plant

+++ include/datafit.h
@@ -17,8 +17,8 @@
 #define currentSlope -10.98
 #define currentZero 38797
 
-uint16_t voltageFit(uint16_t readingValue){};
-uint16_t currentFit(uint16_t readingValue){};
+uint32_t voltageFit(uint16_t readingValue){};
+uint32_t currentFit(uint16_t readingValue){};
 
 // [====================================================]
 // [               IMPLEMENTATION (.c)                  ]
@@ -26,11 +26,11 @@ uint16_t currentFit(uint16_t readingValue){};
 
 // Linar regression is used to convert mesurements into real values
 
-uint16_t voltageFit(uint16_t readingValue){
+uint32_t voltageFit(uint16_t readingValue){
     return(voltageSlope*readingValue+voltageZero);
 }
 
-uint16_t currentFit(uint16_t readingValue){
+uint32_t currentFit(uint16_t readingValue){
     return(currentSlope*readingValue+currentZero);
 }
